Additions and changes to initial part of Group-by section
Added paragraph explaining that most operations apply to
sub-datasets, and gave example to illustrate the idea.

Moved up two lines that relate to this point.

Added explanation of effect of grouping parameters on selection of rows
for sub-datasets.  I couldn't figure out how to describe the map from
group names to indexes in the same way, so I made this item last, and
gave it a different explanation.

Fixed one minor typo in first line of Group-by section (added "s"
to "pack").

Fixed minor typos in line about `grouped?` meta tag.

Additions, changes to beginning of Group-by section

Added paragraph explaining that most operations apply to
sub-datasets, and gave examples to illustrate the idea.

Moved up two lines that relate to this point.

Added explanation of effect of grouping parameters on selection
of rows for sub-datasets.  I couldn't figure out how to describe
the map from group names to indexes in the same way, so I made
this item last, and gave it a different explanation.

Fixed small typo in first line of Group-by section (added "s" to
"pack").

Fixed some typos in the line about the :grouped? meta tag.